% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao.R
\name{analise_residual}
\alias{analise_residual}
\title{Análise Residual}
\usage{
analise_residual(modelo, grafico = TRUE)
}
\arguments{
\item{modelo}{Objeto do tipo lm}

\item{grafico}{Lógico. Se TRUE, exibe o painel 2x2 de diagnóstico dos resíduos.}
}
\description{
Realiza os testes estatísticos de diagnóstico dos resíduos e plota o painel gráfico internamente.
}
