% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{analise_residual_beta}
\alias{analise_residual_beta}
\title{Analise Residual para Modelo Beta}
\usage{
analise_residual_beta(modelo, grafico = TRUE)
}
\arguments{
\item{modelo}{Objeto betareg}

\item{grafico}{Logico. Plota o painel se TRUE.}
}
\description{
Realiza o diagnostico de residuos para regressao Beta (residuos padronizados ponderados).
}
