% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_negativa.R
\name{analise_residual_binomial_negativa}
\alias{analise_residual_binomial_negativa}
\title{Analise Residual para Modelo Binomial Negativo}
\usage{
analise_residual_binomial_negativa(modelo, grafico = TRUE)
}
\arguments{
\item{modelo}{Objeto glm.nb (negbin)}

\item{grafico}{Logico. Plota o painel se TRUE.}
}
\description{
Realiza o diagnostico de residuos para modelos de regressao Binomial Negativa.
}
