% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gama.R
\name{analise_residual_gama}
\alias{analise_residual_gama}
\title{Analise Residual para Modelo Gama}
\usage{
analise_residual_gama(modelo, grafico = TRUE)
}
\arguments{
\item{modelo}{Objeto glm Gama}

\item{grafico}{Logico. Plota o painel se TRUE.}
}
\description{
Painel de residuos e analise de pontos influentes.
}
