% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{analise_residual_poisson}
\alias{analise_residual_poisson}
\title{Analise Residual para Modelo de Poisson}
\usage{
analise_residual_poisson(modelo, grafico = TRUE)
}
\arguments{
\item{modelo}{Objeto glm Poisson}

\item{grafico}{Logico. Plota o painel se TRUE.}
}
\description{
Realiza o diagnostico de residuos para modelos de Poisson.
}
