% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostico.R
\name{diagnostico}
\alias{diagnostico}
\title{Diagnóstico Completo do Banco de Dados}
\usage{
diagnostico(dados)
}
\arguments{
\item{dados}{Data frame ou vetor para análise.}
}
\value{
Retorna o objeto original invisivelmente.
}
\description{
Gera um painel completo mesclando informações de estrutura,
valores ausentes, outliers e valores únicos, alertando para duplicatas.
}
