% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_stat.R
\name{e_stat}
\alias{e_stat}
\title{Create e-Stat object}
\description{
Create e-Stat object

Create e-Stat object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{e_stat$new()}}
\item \href{#method-get_data}{\code{e_stat$get_data()}}
\item \href{#method-restore_key}{\code{e_stat$restore_key()}}
\item \href{#method-restore_value}{\code{e_stat$restore_value()}}
\item \href{#method-clone}{\code{e_stat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{e_stat$new(
  statsDataId,
  appId,
  query = list(),
  url = "http://api.e-stat.go.jp/rest/3.0/app/json/getStatsData"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statsDataId}}{e-Stat statsDataId}

\item{\code{appId}}{Your own e-Stat appId}

\item{\code{query}}{query}

\item{\code{url}}{url}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{e_stat$get_data(limit = 10^5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restore_key"></a>}}
\if{latex}{\out{\hypertarget{method-restore_key}{}}}
\subsection{Method \code{restore_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{e_stat$restore_key()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restore_value"></a>}}
\if{latex}{\out{\hypertarget{method-restore_value}{}}}
\subsection{Method \code{restore_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{e_stat$restore_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{e_stat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
