% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{estat_api_key_exists}
\alias{estat_api_key_exists}
\title{Is an e-Stat API key available?}
\usage{
estat_api_key_exists()
}
\value{
\code{TRUE} if a non-empty \code{ESTAT_API_KEY} is set, otherwise \code{FALSE}.
}
\description{
Is an e-Stat API key available?
}
