% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostico.R
\name{estrutura}
\alias{estrutura}
\title{Estrutura do Banco de Dados}
\usage{
estrutura(dados)
}
\arguments{
\item{dados}{Data frame para análise.}
}
\value{
Retorna o data frame invisivelmente.
}
\description{
Retorna apenas os tipos de dados de cada variável.
}
