% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{med_forma}
\alias{med_forma}
\title{Medidas de Forma da Distribuição}
\usage{
med_forma(x, decimais = 2)
}
\arguments{
\item{x}{Vetor numérico.}

\item{decimais}{Casas decimais (padrão 2).}
}
\description{
Medidas de Forma da Distribuição
}
