% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{med_tend_central}
\alias{med_tend_central}
\title{Medidas de Tendência Central}
\usage{
med_tend_central(x, decimais = 2)
}
\arguments{
\item{x}{Vetor numérico.}

\item{decimais}{Casas decimais (padrão 2).}
}
\description{
Medidas de Tendência Central
}
