% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao.R
\name{metricas}
\alias{metricas}
\title{Métricas do Modelo}
\usage{
metricas(modelo, corte_prob = 0.5)
}
\arguments{
\item{modelo}{Objeto do tipo lm}
}
\description{
Exibe as principais métricas de avaliação e erro para um modelo de regressão.
}
