% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{metricas_beta}
\alias{metricas_beta}
\title{Metricas do Modelo Beta}
\usage{
metricas_beta(modelo)
}
\arguments{
\item{modelo}{Objeto de regressao beta (betareg)}
}
\description{
Calcula e exibe as principais metricas para um modelo betareg.
}
