% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_negativa.R
\name{metricas_binomial_negativa}
\alias{metricas_binomial_negativa}
\title{Metricas do Modelo Binomial Negativo}
\usage{
metricas_binomial_negativa(modelo)
}
\arguments{
\item{modelo}{Objeto glm.nb (negbin)}
}
\description{
Calcula e exibe as principais metricas para um modelo glm.nb.
}
