% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gama.R
\name{metricas_gama}
\alias{metricas_gama}
\title{Metricas do Modelo Gama}
\usage{
metricas_gama(modelo)
}
\arguments{
\item{modelo}{Objeto glm Gama}
}
\description{
Calcula as principais metricas para um modelo glm Gama.
}
