% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{metricas_poisson}
\alias{metricas_poisson}
\title{Metricas do Modelo Poisson}
\usage{
metricas_poisson(modelo)
}
\arguments{
\item{modelo}{Objeto glm Poisson}
}
\description{
Calcula e exibe as principais metricas para um modelo glm Poisson.
}
