% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghelpers.R
\name{paleta_estatR}
\alias{paleta_estatR}
\title{Paleta de Cores do estatR}
\usage{
paleta_estatR(nome = NULL)
}
\arguments{
\item{nome}{Nome da paleta (ex: "academic", "vibrant", "nature"). Se NULL, plota todas as paletas disponíveis.}
}
\value{
Um objeto ggplot_add para adicionar ao gráfico, ou (se chamada sem argumentos)
plota um demonstrativo visual de todas as paletas na aba Plots.
}
\description{
Aplica paletas de cores padronizadas aos gráficos ggplot2.
}
