% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{percentis}
\alias{percentis}
\title{Percentis da Distribuição}
\usage{
percentis(x, p = c(5, 10, 25, 50, 75, 90, 95), decimais = 2)
}
\arguments{
\item{x}{Vetor numérico.}

\item{p}{Vetor numérico com os percentis desejados (entre 0 e 100). Padrão: c(5, 10, 25, 50, 75, 90, 95).}

\item{decimais}{Casas decimais (padrão 2).}
}
\value{
Retorna invisivelmente um vetor nomeado com os valores dos percentis.
}
\description{
Calcula e exibe os percentis de um vetor numérico. Por padrão exibe
P5, P10, P25, P50, P75, P90 e P95, mas o usuário pode definir quais
percentis quer calcular pelo parâmetro \code{p}.
}
