% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilidade.R
\name{prob_binomial}
\alias{prob_binomial}
\title{Probabilidade da Distribuição Binomial}
\usage{
prob_binomial(
  ensaios,
  prob,
  q,
  tipo = c("exato", "menor", "maior"),
  grafico = TRUE
)
}
\arguments{
\item{ensaios}{Número de tentativas (n).}

\item{prob}{Probabilidade de sucesso (p).}

\item{q}{Valor de interesse (número de sucessos).}

\item{tipo}{"exato" (P(X=q)), "menor" (P(X<=q)), "maior" (P(X>=q)).}

\item{grafico}{Lógico. Se TRUE, gera o gráfico.}
}
\description{
Calcula a probabilidade para uma distribuição Binomial e plota as barras.
}
