% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilidade.R
\name{prob_exponencial}
\alias{prob_exponencial}
\title{Probabilidade da Distribuição Exponencial}
\usage{
prob_exponencial(
  taxa = 1,
  q1 = NULL,
  q2 = NULL,
  tipo = c("menor", "maior", "entre"),
  grafico = TRUE
)
}
\arguments{
\item{taxa}{Taxa de ocorrência (lambda). Deve ser positiva.}

\item{q1}{Quantil de referência.}

\item{q2}{Segundo quantil (apenas para tipo = "entre").}

\item{tipo}{Tipo: "menor" P(X < q1), "maior" P(X > q1), "entre" P(q1 < X < q2).}

\item{grafico}{Lógico. Se TRUE, plota a curva com área sombreada.}
}
\description{
Calcula a probabilidade para uma distribuição Exponencial e plota a curva com área sombreada.
}
