% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilidade.R
\name{prob_normal}
\alias{prob_normal}
\title{Probabilidade da Distribuição Normal}
\usage{
prob_normal(
  media = 0,
  dp = 1,
  q1 = NULL,
  q2 = NULL,
  tipo = c("menor", "maior", "entre"),
  grafico = TRUE
)
}
\arguments{
\item{media}{Média teórica da distribuição.}

\item{dp}{Desvio-padrão teórico.}

\item{q1}{Quantil de referência (limite de integração).}

\item{q2}{Segundo quantil de referência (apenas para tipo = "entre").}

\item{tipo}{Tipo de probabilidade desejada: "menor" (P(X < q1)), "maior" (P(X > q1)) ou "entre" (P(q1 < X < q2)).}

\item{grafico}{Lógico. Se TRUE, gera o gráfico da distribuição com a área hachurada.}
}
\description{
Calcula a probabilidade para uma distribuição Normal e gera o gráfico com a área sombreada.
}
