% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilidade.R
\name{prob_poisson}
\alias{prob_poisson}
\title{Probabilidade da Distribuição Poisson}
\usage{
prob_poisson(lambda, q, tipo = c("exato", "menor", "maior"), grafico = TRUE)
}
\arguments{
\item{lambda}{Taxa de ocorrência.}

\item{q}{Valor de interesse.}

\item{tipo}{"exato" (P(X=q)), "menor" (P(X<=q)), "maior" (P(X>=q)).}

\item{grafico}{Lógico. Se TRUE, gera o gráfico.}
}
\description{
Calcula a probabilidade para uma distribuição Poisson e plota as barras.
}
