% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{quintis}
\alias{quintis}
\title{Quintis da Distribuição}
\usage{
quintis(x, decimais = 2)
}
\arguments{
\item{x}{Vetor numérico.}

\item{decimais}{Casas decimais (padrão 2).}
}
\description{
Quintis da Distribuição
}
