% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{regressao_beta}
\alias{regressao_beta}
\title{Regressao Beta}
\usage{
regressao_beta(formula, dados, grafico = TRUE)
}
\arguments{
\item{formula}{Formula do modelo (ex: Y ~ X1 + X2).}

\item{dados}{Data frame contendo as variaveis.}

\item{grafico}{Logico. Se TRUE, plota o grafico de Observado vs Predito.}
}
\description{
Ajusta um modelo de regressao Beta para variaveis respostas
que sao proporcoes ou taxas estritamente no intervalo (0, 1).
Apresenta estatisticas de ajuste, parametro de precisao (phi), coeficientes,
Odds Ratio (OR) para a media e interpretacao automatizada.
}
