% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_negativa.R
\name{regressao_binomial_negativa}
\alias{regressao_binomial_negativa}
\title{Regressao Binomial Negativa}
\usage{
regressao_binomial_negativa(formula, dados, grafico = TRUE)
}
\arguments{
\item{formula}{Formula do modelo (ex: Y ~ X1 + X2).}

\item{dados}{Data frame contendo as variaveis.}

\item{grafico}{Logico. Se TRUE, plota o grafico de Observado vs Predito.}
}
\description{
Ajusta um modelo de regressao Binomial Negativa para dados de contagem
com sobredispersao (variancia > media). Apresenta estatisticas de ajuste,
parametro de dispersao theta, coeficientes, Razao de Taxa de Incidencia (IRR)
e interpretacao automatizada.
}
