% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gama.R
\name{regressao_gama}
\alias{regressao_gama}
\title{Regressao Gama}
\usage{
regressao_gama(formula, dados, grafico = TRUE)
}
\arguments{
\item{formula}{Formula do modelo (ex: Y ~ X1 + X2).}

\item{dados}{Data frame contendo as variaveis.}

\item{grafico}{Logico. Se TRUE, plota o grafico de Observado vs Predito.}
}
\description{
Ajusta um modelo de regressao Gama (link = "log") para variaveis
contínuas e estritamente positivas, geralmente com forte assimetria à direita.
Apresenta estatisticas de ajuste, parametro de forma, coeficientes (como multiplicadores da média)
e interpretacao automatizada.
}
