% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao.R
\name{regressao_linear}
\alias{regressao_linear}
\title{Regressão Linear Simples e Múltipla}
\usage{
regressao_linear(formula, dados, grafico = TRUE, ...)
}
\arguments{
\item{formula}{Fórmula do modelo (ex: y ~ x)}

\item{dados}{Data frame com os dados}

\item{grafico}{Lógico. Se TRUE, gera um gráfico de dispersão para regressão simples.}

\item{...}{Outros argumentos passados para o gráfico.}
}
\description{
Ajusta um modelo de regressão linear e retorna uma saída formatada com coeficientes, ajuste e interpretação.
}
