% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{regressao_poisson}
\alias{regressao_poisson}
\title{Regressao de Poisson}
\usage{
regressao_poisson(formula, dados, grafico = TRUE)
}
\arguments{
\item{formula}{Formula do modelo (ex: Y ~ X1 + X2).}

\item{dados}{Data frame contendo as variaveis.}

\item{grafico}{Logico. Se TRUE, plota o grafico de Observado vs Predito.}
}
\description{
Ajusta um modelo de regressao de Poisson para dados de contagem.
Apresenta estatisticas de ajuste, fator de sobredispersao, coeficientes,
Razao de Taxa de Incidencia (IRR) e interpretacao automatizada.
}
