% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{selecao_modelos_beta}
\alias{selecao_modelos_beta}
\title{Selecao de Modelos - Beta}
\usage{
selecao_modelos_beta(formula, dados, top = 5)
}
\arguments{
\item{formula}{Formula do modelo}

\item{dados}{Data frame}

\item{top}{Numero de top modelos a retornar}
}
\description{
Realiza Best Subsets para regressao Beta.
}
