% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_negativa.R
\name{selecao_modelos_binomial_negativa}
\alias{selecao_modelos_binomial_negativa}
\title{Selecao de Modelos - Binomial Negativa}
\usage{
selecao_modelos_binomial_negativa(formula, dados, top = 5)
}
\arguments{
\item{formula}{Formula do modelo}

\item{dados}{Data frame}

\item{top}{Numero de top modelos a retornar}
}
\description{
Realiza Best Subsets para regressao Binomial Negativa.
}
