% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gama.R
\name{selecao_modelos_gama}
\alias{selecao_modelos_gama}
\title{Selecao de Modelos - Gama}
\usage{
selecao_modelos_gama(formula, dados, top = 5)
}
\arguments{
\item{formula}{Formula do modelo}

\item{dados}{Data frame}

\item{top}{Numero de top modelos a retornar}
}
\description{
Realiza Best Subsets para regressao Gama.
}
