% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{selecao_modelos_poisson}
\alias{selecao_modelos_poisson}
\title{Selecao de Modelos - Poisson}
\usage{
selecao_modelos_poisson(formula, dados, top = 5)
}
\arguments{
\item{formula}{Formula do modelo}

\item{dados}{Data frame}

\item{top}{Numero de top modelos a retornar}
}
\description{
Realiza Best Subsets para regressao de Poisson.
}
