% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostico.R
\name{valores_ausentes}
\alias{valores_ausentes}
\title{Análise de Valores Ausentes}
\usage{
valores_ausentes(dados)
}
\arguments{
\item{dados}{Data frame ou vetor para análise.}
}
\value{
Retorna o objeto original invisivelmente.
}
\description{
Inspeciona procurando por NAs oficiais e valores
que indicam respostas vazias ou erros ocultos (como -99 ou textos vazios).
}
