% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{descrever}
\alias{descrever}
\title{Resumo Estatistico de Variaveis ou Banco de Dados}
\usage{
descrever(x, por = NULL, numericas = TRUE, categoricas = TRUE, decimais = 2)
}
\arguments{
\item{x}{Um data frame ou vetor.}

\item{por}{Variavel categorica para agrupar o resumo.}

\item{numericas}{Logico. Se TRUE, exibe numericas.}

\item{categoricas}{Logico. Se TRUE, exibe categoricas.}

\item{decimais}{Casas decimais (padrao 2).}
}
\description{
Gera um resumo estatistico completo.
}
