% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostico.R
\name{diagnostico}
\alias{diagnostico}
\title{Diagnóstico Completo do Banco de Dados}
\usage{
diagnostico(dados)
}
\arguments{
\item{dados}{Data frame para análise.}
}
\value{
Retorna o data frame invisivelmente.
}
\description{
Gera um painel completo mesclando informações de estrutura e
valores ausentes por tipo de variável, incluindo detecção automática de
NAs ocultos (como -99 ou vazio).
}
