% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{diagnostico}
\alias{diagnostico}
\title{Diagnóstico de Dados ou Variáveis}
\usage{
diagnostico(x)
}
\arguments{
\item{x}{Um data.frame ou um vetor atômico (numérico, categórico, etc.).}
}
\value{
Retorna o objeto invisivelmente.
}
\description{
Exibe uma visão geral rápida do banco de dados ou de uma variável específica,
mostrando dimensões, contagem de válidos/NAs e proporções.
}
